requirements.txt
# Core — needs 1.31+ for st.write_stream()
streamlit>=1.31.0
streamlit-option-menu==0.3.6

# HTTP
requests==2.31.0
httpx==0.25.0

# Data
pandas
numpy

# Visualisation
plotly
matplotlib
seaborn
altair

# Image / media
Pillow
opencv-python

# Config & env
python-dotenv
pydantic>=2.7.4
pydantic-settings

# Cache
diskcache

# Date / text
python-dateutil
pytz
beautifulsoup4
markdown

# UI extras
stqdm
streamlit-lottie
streamlit-card

# Logging
loguru

# AI — Groq (llama-3.3-70b-versatile) with streaming
groq
langchain>=0.3.27
langchain-groq
langchain-core

# Dev / QA (optional)
pytest
black
flake8
isort
mypy
types-requests

# Misc
click
tqdm
urllib3
streamlit-aggrid
